+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Change Log for SD.Next
 
-## Update for 2024-04-07
+## Update for 2024-04-09
 
 - **Features**:
   - **Gallery**: list, preview, search through all your images and videos!  
@@ -33,6 +33,7 @@
   - Faster server startup  
   - Add **MIGraphX** torch optimization engine, thanks @Disty0
   - Styles apply wildcards to params
+  - Add option to make batch generations use fully random seed vs sequential  
   - Make metadata in full screen viewer optional
   - Add VAE civitai scan metadata/preview
   - More efficient in-browser callbacks